unity-backend/requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
cached-property==1.5.2
celery==5.2.3
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
Django==3.2.16
django-celery-results==2.4.0
django-cors-headers==3.13.0
django-rest-framework==0.1.0
djangorestframework==3.14.0
humanfriendly==10.0
humanize==4.4.0
importlib-metadata==4.13.0
kombu==5.2.4
packaging==21.3
prompt-toolkit==3.0.31
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
redis==4.3.4
six==1.16.0
sqlparse==0.4.3
typing-extensions==4.4.0
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
zipp==3.10.0
